requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.7
django-bootstrap5==23.4
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.49.0
frozendict==2.4.0
GetOldTweets3==0.0.11
html5lib==1.1
httpie==3.2.2
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
Pygments==2.16.1
pyparsing==3.1.1
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
xlrd==2.0.1
yfinance==0.2.37
